+++ styles/onboarding.module.css
@@ -1,0 +1,24 @@
+
+.flexRow {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 8px;
+}
+
+.container {
+    display: flex;
+    flex-direction: column;
+    padding: 16px;
+    gap: 4px;
+}
+
+.githubButton {
+    width: 24px;
+    height: 24px;
+    background: transparent;
+    border: none;
+}
+.githubButton:hover {
+    cursor: pointer;
+}
