+++ utils/functions/getCurrentYear.ts
@@ -1,0 +1,4 @@
+export default function getCurrentYear() {
+    const currentYear = new Date().getFullYear()
+    return currentYear || 2024
+}
